+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.6.2]
+
+* Fix wrong selection offset
+* Fix wrong text clip due to overflowspan
+
 ## [0.6.1]
 
 * Fix issue [type 'List<InlineSpan>' is not a subtype of type 'List<TextSpan>'](https://github.com/fluttercandies/extended_text/issues/20)